+++ src/Resources/EmbeddedResources/Message.php
@@ -1,0 +1,78 @@
+<?php
+/**
+ * Represents the message resource holding information like a transactions error code and message.
+ *
+ * Copyright (C) 2018 heidelpay GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @link  http://dev.heidelpay.com/
+ *
+ * @author  Simon Gabriel <development@heidelpay.com>
+ *
+ * @package  heidelpayPHP/resources/embedded_resources
+ */
+namespace heidelpayPHP\Resources\EmbeddedResources;
+
+use heidelpayPHP\Resources\AbstractHeidelpayResource;
+
+class Message extends AbstractHeidelpayResource
+{
+    /** @var string $code */
+    private $code = '';
+
+    /** @var string $customer */
+    private $customer = '';
+
+    //<editor-fold desc="Getters/Setters">
+
+    /**
+     * @return string
+     */
+    public function getCode(): string
+    {
+        return $this->code;
+    }
+
+    /**
+     * @param string $code
+     *
+     * @return Message
+     */
+    public function setCode(string $code): Message
+    {
+        $this->code = $code;
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getCustomer(): string
+    {
+        return $this->customer;
+    }
+
+    /**
+     * @param string $customer
+     *
+     * @return Message
+     */
+    public function setCustomer(string $customer): Message
+    {
+        $this->customer = $customer;
+        return $this;
+    }
+
+    //</editor-fold>
+}